requirements.txt
pandas>=2.0
numpy>=1.24
tqdm>=4.66
pyyaml>=6.0
requests>=2.31
beautifulsoup4>=4.12
nltk>=3.9
spacy>=3.7
gensim>=4.3
sentence-transformers>=3.0
scikit-learn>=1.4
bertopic>=0.16
umap-learn>=0.5.6
hdbscan>=0.8.36
faiss-cpu>=1.8.0
joblib>=1.4
snscrape==0.7.0.20230622
newspaper3k>=0.2.8
transformers>=4.42
torch>=2.2
dowhy>=0.11
econml>=0.15
streamlit>=1.36
pytest>=8.2
pyarrow>=16.1
